exec() の match アーム本体を関数へ抽出する #96
exec() 本体 (840行) の match アームのうち、自己完結しているものを
`fn exec_xxx(song: &mut Song, t: &Token)` として抽出し、
`TokenType::X => exec_x(song, t),` の形に揃えた。

新規サブモジュール:
- runner/sysex.rs       SysEx (GSEffect / SysExCommand / リセット等)
- runner/meta.rs        メタイベント・ログ出力・SMFへの直接出力
- runner/variable.rs    変数の定義・参照と計算木の実行
- runner/track_state.rs トラックの演奏パラメータ (l/o/v/q/t) の設定

既存サブモジュールへの追加:
- runner/cc.rs        CC・ピッチベンド・テンポ関連
- runner/structure.rs トラック/チャンネル/キー等の曲構造
- runner/control.rs   Return の戻り値設定

exec() に残したのは LoopBegin / LoopBreak / LoopEnd の3アームのみ。
これらは pos と loop_stack を直接操作するため抽出できない。
Issue で挙げられていた他の10アームは、`pos += 1; continue;` が
フォールスルーと等価だったため `return` に置き換えて抽出した。
Break / Continue / Return は while ループの `break` が必要なため
本体のみを抽出し、`break` は exec() に残している。

VelocityOnNote/OnCycle など on-note と on-cycle の対になるアームは
is_cycle 引数を持つ1つの関数にまとめた (振る舞いは同一)。

結果: exec() 840行 → 203行、runner.rs 981行 → 324行

振る舞いは変更していない。cargo test 102件通過、ゴールデンMIDIも一致。
さらに main のバイナリと比較し、samples/ 全曲と主要コマンドを網羅した
27本のMMLでMIDIバイナリ・ログ出力が完全一致することを確認した
(o.Random 等の乱数系は元から実行ごとに出力が変わるため対象外)。
